[refactor] lost-animal 디렉토리 패키지 구조 변경
