+++ src/main/java/kr/co/pawong/pwbe/lostAnimal/adapter/in/api/LostAnimalQueryService.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostAnimal.adapter.in.api;
+
+public interface LostAnimalQueryService {
+
+}
