--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/domain/LostAnimal.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.domain;
-
-public class LostAnimal {
-
-}
