--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/domain/LostAnimalCreate.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.domain;
-
-public class LostAnimalCreate {
-
-}
