+++ src/main/java/kr/co/pawong/pwbe/lostAnimal/application/service/LostAnimalUpdateService.java
@@ -1,5 +1,5 @@
 package kr.co.pawong.pwbe.lostAnimal.application.service;
 
-public class LostAnimalUpdateServiceImpl {
+public class LostAnimalUpdateService {
 
 }
