--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/service/port/LostAnimalQueryRepository.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.service.port;
-
-public interface LostAnimalQueryRepository {
-
-}
