+++ src/main/java/kr/co/pawong/pwbe/lostAnimal/domain/LostAnimalCreate.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostAnimal.domain;
+
+public class LostAnimalCreate {
+
+}
