--- src/main/java/kr/co/pawong/pwbe/lostAnimal/infrastructure/repository/LostAnimalJpaRepository.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.infrastructure.repository;
-
-public class LostAnimalJpaRepository {
-
-}
