--- src/main/java/kr/co/pawong/pwbe/lostAnimal/presentation/controller/LostAnimalUpdateController.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.presentation.controller;
-
-public class LostAnimalUpdateController {
-
-}
