--- src/main/java/kr/co/pawong/pwbe/lostAnimal/presentation/controller/port/LostAnimalQueryService.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.presentation.controller.port;
-
-public interface LostAnimalQueryService {
-
-}
